Move get and apply to definition manager
