+++ src/Microsoft.Health.Fhir.Core/Features/Definition/SearchParameterDefinitionManager.cs
@@ -36,7 +36,7 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
     /// <summary>
     /// Provides mechanism to access search parameter definition.
     /// </summary>
-    public class SearchParameterDefinitionManager : ISearchParameterDefinitionManager, INotificationHandler<SearchParametersUpdatedNotification>, INotificationHandler<StorageInitializedNotification>
+    public class SearchParameterDefinitionManager : ISearchParameterDefinitionManager, INotificationHandler<SearchParametersUpdatedNotification>, INotificationHandler<StorageInitializedNotification>, IDisposable
     {
         private readonly IModelInfoProvider _modelInfoProvider;
         private readonly IMediator _mediator;
@@ -44,8 +44,11 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
         private readonly IScopeProvider<ISearchService> _searchServiceFactory;
         private readonly ISearchParameterComparer<SearchParameterInfo> _searchParameterComparer;
         private readonly IScopeProvider<ISearchParameterStatusDataStore> _searchParameterStatusDataStoreFactory;
-        private readonly IScopeProvider<IFhirDataStore> _fhirDataStoreFactory;
+        private ISearchParameterStatusDataStore _statusDataStore;
         private readonly ILogger _logger;
+        private DateTimeOffset? _searchParamLastUpdated;
+        private readonly SemaphoreSlim _refreshSemaphore;
+        private readonly int maxSecondsToWait = 100;
 
         private bool _initialized = false;
 
@@ -55,7 +58,6 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
             IScopeProvider<ISearchService> searchServiceFactory,
             ISearchParameterComparer<SearchParameterInfo> searchParameterComparer,
             IScopeProvider<ISearchParameterStatusDataStore> searchParameterStatusDataStoreFactory,
-            IScopeProvider<IFhirDataStore> fhirDataStoreFactory,
             ILogger<SearchParameterDefinitionManager> logger)
         {
             EnsureArg.IsNotNull(modelInfoProvider, nameof(modelInfoProvider));
@@ -63,7 +65,6 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
             EnsureArg.IsNotNull(searchServiceFactory, nameof(searchServiceFactory));
             EnsureArg.IsNotNull(searchParameterComparer, nameof(searchParameterComparer));
             EnsureArg.IsNotNull(searchParameterStatusDataStoreFactory, nameof(searchParameterStatusDataStoreFactory));
-            EnsureArg.IsNotNull(fhirDataStoreFactory, nameof(fhirDataStoreFactory));
             EnsureArg.IsNotNull(logger, nameof(logger));
 
             _modelInfoProvider = modelInfoProvider;
@@ -74,8 +75,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
             _searchServiceFactory = searchServiceFactory;
             _searchParameterComparer = searchParameterComparer;
             _searchParameterStatusDataStoreFactory = searchParameterStatusDataStoreFactory;
-            _fhirDataStoreFactory = fhirDataStoreFactory;
             _logger = logger;
+            _refreshSemaphore = new SemaphoreSlim(1, 1);
 
             var bundle = SearchParameterDefinitionBuilder.ReadEmbeddedSearchParameters("search-parameters.json", _modelInfoProvider);
             var msBundle = SearchParameterDefinitionBuilder.ReadEmbeddedSearchParameters("ms-search-parameters.json", _modelInfoProvider);
@@ -93,6 +94,8 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
                 isSystemDefined: true);
         }
 
+        public DateTimeOffset? SearchParamLastUpdated => _searchParamLastUpdated;
+
         internal ConcurrentDictionary<string, SearchParameterInfo> UrlLookup { get; set; }
 
         // TypeLookup: Resource type -> code -> ordered queue of definition URLs.
@@ -109,10 +112,10 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
         {
             try
             {
-                _initialized = true;
-                await LoadSearchParamsFromDataStore(cancellationToken);
-
+                _statusDataStore = _searchParameterStatusDataStoreFactory.Invoke().Value;
+                await GetAndApplySearchParameterUpdates(cancellationToken);
                 await _mediator.Publish(new SearchParameterDefinitionManagerInitialized(), cancellationToken);
+                _initialized = true;
             }
             catch
             {
@@ -369,190 +372,6 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
             // await _mediator.Publish(new ImproperBehaviorNotification("Error initializing search parameters"), cancellationToken);
         }
 
-        private async Task LoadSearchParamsFromDataStore(CancellationToken cancellationToken)
-        {
-            // now read in any previously POST'd SearchParameter resources
-            using IScoped<ISearchService> search = _searchServiceFactory.Invoke();
-            using IScoped<ISearchParameterStatusDataStore> statusDataStore = _searchParameterStatusDataStoreFactory.Invoke();
-            using IScoped<IFhirDataStore> fhirDataStore = _fhirDataStoreFactory.Invoke();
-
-            string continuationToken = null;
-            int totalLoaded = 0;
-            int totalPendingDelete = 0;
-
-            // Get all PendingDelete search parameters from the status store
-            var allStatuses = await statusDataStore.Value.GetSearchParameterStatuses(cancellationToken);
-            var pendingDeleteUrls = new HashSet<string>(
-                allStatuses
-                    .Where(s => s.Status == SearchParameterStatus.PendingDelete)
-                    .Select(s => s.Uri.OriginalString),
-                StringComparer.Ordinal);
-
-            _logger.LogInformation(
-                "Found {PendingDeleteCount} search parameters with PendingDelete status in the status store",
-                pendingDeleteUrls.Count);
-
-            do
-            {
-                var searchOptions = new SearchOptions();
-                searchOptions.Sort = new List<(SearchParameterInfo, SortOrder)>();
-                searchOptions.UnsupportedSearchParams = new List<Tuple<string, string>>();
-                searchOptions.Expression = Expression.SearchParameter(
-                    SearchParameterInfo.ResourceTypeSearchParameter,
-                    Expression.StringEquals(FieldName.TokenCode, null, KnownResourceTypes.SearchParameter, false));
-                searchOptions.MaxItemCount = 10;
-
-                // ✅ Include soft-deleted resources to find PendingDelete search parameters
-                searchOptions.ResourceVersionTypes = ResourceVersionType.Latest | ResourceVersionType.SoftDeleted;
-
-                if (continuationToken != null)
-                {
-                    searchOptions.ContinuationToken = continuationToken;
-                }
-
-                var result = await search.Value.SearchAsync(searchOptions, cancellationToken);
-                continuationToken = result?.ContinuationToken;
-
-                if (result?.Results != null && result.Results.Any())
-                {
-                    foreach (var searchResult in result.Results)
-                    {
-                        var isDeleted = searchResult.Resource.IsDeleted;
-
-                        // For soft-deleted resources, check if they are in PendingDelete status
-                        if (isDeleted)
-                        {
-                            try
-                            {
-                                // Get the resource ID to fetch its last version before deletion
-                                var resourceId = searchResult.Resource.ResourceId;
-
-                                // Parse the current version and calculate the previous version
-                                if (int.TryParse(searchResult.Resource.Version, out int currentVersion) && currentVersion > 1)
-                                {
-                                    var previousVersion = (currentVersion - 1).ToString();
-                                    var resourceKey = new ResourceKey(KnownResourceTypes.SearchParameter, resourceId, previousVersion);
-                                    var lastVersion = await fhirDataStore.Value.GetAsync(resourceKey, cancellationToken);
-
-                                    if (lastVersion?.RawResource != null)
-                                    {
-                                        var searchParam = lastVersion.RawResource.ToITypedElement(_modelInfoProvider);
-                                        var urlScalar = searchParam.GetStringScalar("url");
-
-                                        // Only load if this URL is marked as PendingDelete in the status store
-                                        if (!string.IsNullOrEmpty(urlScalar) && pendingDeleteUrls.Contains(urlScalar))
-                                        {
-                                            // Build the search parameter using the last version before deletion
-                                            SearchParameterDefinitionBuilder.Build(
-                                                new List<ITypedElement>() { searchParam },
-                                                UrlLookup,
-                                                TypeLookup,
-                                                _modelInfoProvider,
-                                                _searchParameterComparer,
-                                                _logger);
-
-                                            totalLoaded++;
-
-                                            // Update the status to PendingDelete since the resource is soft-deleted
-                                            if (UrlLookup.TryGetValue(urlScalar, out var loadedParam))
-                                            {
-                                                loadedParam.SearchParameterStatus = SearchParameterStatus.PendingDelete;
-                                                totalPendingDelete++;
-                                                _logger.LogInformation(
-                                                    "Loaded PendingDelete search parameter from last version before deletion: {Url}",
-                                                    urlScalar);
-                                            }
-                                        }
-                                        else if (!string.IsNullOrEmpty(urlScalar))
-                                        {
-                                            _logger.LogDebug(
-                                                "Skipping soft-deleted SearchParameter {ResourceId} with URL {Url} - not in PendingDelete status",
-                                                resourceId,
-                                                urlScalar);
-                                        }
-                                        else
-                                        {
-                                            _logger.LogWarning(
-                                                "Could not retrieve valid URL for soft-deleted SearchParameter {ResourceId}",
-                                                resourceId);
-                                        }
-                                    }
-                                    else
-                                    {
-                                        _logger.LogWarning(
-                                            "Could not retrieve last version for soft-deleted SearchParameter {ResourceId}",
-                                            resourceId);
-                                    }
-                                }
-                                else
-                                {
-                                    _logger.LogWarning(
-                                        "Could not parse version or version is 1 for soft-deleted SearchParameter {ResourceId}, version: {Version}",
-                                        resourceId,
-                                        searchResult.Resource.Version);
-                                }
-                            }
-                            catch (Exception ex)
-                            {
-                                _logger.LogError(
-                                    ex,
-                                    "Error loading last version of soft-deleted SearchParameter {ResourceId}",
-                                    searchResult.Resource.ResourceId);
-                            }
-                        }
-                        else
-                        {
-                            // Normal processing for active resources
-                            var searchParam = searchResult.Resource.RawResource.ToITypedElement(_modelInfoProvider);
-
-                            try
-                            {
-                                SearchParameterDefinitionBuilder.Build(
-                                    new List<ITypedElement>() { searchParam },
-                                    UrlLookup,
-                                    TypeLookup,
-                                    _modelInfoProvider,
-                                    _searchParameterComparer,
-                                    _logger);
-
-                                totalLoaded++;
-                            }
-                            catch (FhirException ex)
-                            {
-                                StringBuilder issueDetails = new StringBuilder();
-                                foreach (OperationOutcomeIssue issue in ex.Issues)
-                                {
-                                    issueDetails.Append(issue.Diagnostics).Append("; ");
-                                }
-
-                                _logger.LogWarning(
-                                    ex,
-                                    "Error loading search parameter {Url} from data store. Issues: {Issues}",
-                                    searchParam.GetStringScalar("url"),
-                                    issueDetails.ToString());
-                            }
-                            catch (Exception ex) when (
-                                !(ex is OutOfMemoryException
-                                || ex is StackOverflowException
-                                || ex is ThreadAbortException))
-                            {
-                                _logger.LogError(
-                                    ex,
-                                    "Error loading search parameter {Url} from data store.",
-                                    searchParam.GetStringScalar("url"));
-                            }
-                        }
-                    }
-                }
-            }
-            while (continuationToken != null);
-
-            _logger.LogInformation(
-                "Loaded {TotalLoaded} active and {TotalPendingDelete} PendingDelete search parameters from data store",
-                totalLoaded,
-                totalPendingDelete);
-        }
-
         private bool TryGetFromTypeLookup(string resourceType, string code, out SearchParameterInfo searchParameter)
         {
             searchParameter = null;
@@ -605,6 +424,295 @@ namespace Microsoft.Health.Fhir.Core.Features.Definition
             }
 
             return completeResourceList;
+        }
+
+        /// <summary>
+        /// This method should be called to get any updates to search params cache.
+        /// </summary>
+        /// <param name="cancellationToken">Cancellation token</param>
+        /// <param name="zeroWaitForSemaphore">Whether to wait for the semaphore to become available.</param>
+        /// <returns>A task that returns true if the refresh was performed, false if it was skipped due to exceeding the lock interval.</returns>
+        public async Task<bool> GetAndApplySearchParameterUpdates(CancellationToken cancellationToken = default, bool zeroWaitForSemaphore = false)
+        {
+            if (!await _refreshSemaphore.WaitAsync(TimeSpan.FromSeconds(zeroWaitForSemaphore ? 0 : maxSecondsToWait), cancellationToken))
+            {
+                var msg = $"Could not acquire lock to refresh search parameter cache after waiting for {maxSecondsToWait} seconds.";
+                if (zeroWaitForSemaphore)
+                {
+                    _logger.LogInformation(msg);
+                    return false;
+                }
+                else
+                {
+                    _logger.LogError(msg);
+                    throw new InvalidOperationException(msg);
+                }
+            }
+
+            try
+            {
+                var results = await GetSearchParameterStatusUpdates(cancellationToken, _searchParamLastUpdated);
+                var statuses = results.Statuses;
+
+                // First process any deletes or disables, then we will do any adds or updates
+                // this way any deleted or params which might have the same code or name as a new
+                // parameter will not cause conflicts. Disabled params just need to be removed when calculating the hash.
+                foreach (var searchParam in statuses.Where(p => p.Status == SearchParameterStatus.Deleted))
+                {
+                    DeleteSearchParameterIfExists(searchParam.Uri.OriginalString);
+                }
+
+                foreach (var searchParam in statuses.Where(p => p.Status == SearchParameterStatus.PendingDelete))
+                {
+                    UpdateSearchParameterStatus(searchParam.Uri.OriginalString, SearchParameterStatus.PendingDelete);
+                }
+
+                // Identify all System Defined Search Parameters and filter them from statuses
+                var systemDefinedSearchParameterUris = new HashSet<string>(AllSearchParameters.Where(p => p.IsSystemDefined).Select(p => p.Url.OriginalString), StringComparer.Ordinal);
+
+                var statusesToFetch = statuses
+                    .Where(p => p.Status == SearchParameterStatus.Enabled || p.Status == SearchParameterStatus.Supported)
+                    .Where(p => !systemDefinedSearchParameterUris.Contains(p.Uri.OriginalString)).ToList();
+
+                // Batch fetch all SearchParameter resources in one call
+                var searchParamResources = await GetSearchParametersByUrls(
+                    statusesToFetch
+                        .Select(p => p.Uri.OriginalString)
+                        .ToList(),
+                    cancellationToken);
+
+                var paramsToAdd = new List<ITypedElement>();
+                var allHaveResources = true;
+                foreach (var searchParam in statusesToFetch)
+                {
+                    if (!searchParamResources.TryGetValue(searchParam.Uri.OriginalString, out var searchParamResource))
+                    {
+                        _logger.LogInformation(
+                            "Updated SearchParameter status found for SearchParameter: {Url}, but did not find any SearchParameter resources when querying for this url.",
+                            searchParam.Uri);
+
+                        if (searchParam.LastUpdated > DateTimeOffset.UtcNow.AddMinutes(-10)) // same as for in cache
+                        {
+                            allHaveResources = false;
+                        }
+
+                        continue;
+                    }
+
+                    // check if search param is in cache and add if does not exist
+                    if (TryGetSearchParameter(searchParam.Uri.OriginalString, out var existingSearchParam))
+                    {
+                        // if the previous version of the search parameter exists we should delete the old information currently stored
+                        DeleteSearchParameter(searchParam.Uri.OriginalString);
+                    }
+
+                    paramsToAdd.Add(searchParamResource);
+
+                    // Add parameters incrementally per chunk to reduce peak memory footprint
+                    if (paramsToAdd.Count >= 100)
+                    {
+                        AddNewSearchParameters(paramsToAdd);
+                        paramsToAdd.Clear();
+                    }
+                }
+
+                // Add any remaining parameters
+                if (paramsToAdd.Any())
+                {
+                    AddNewSearchParameters(paramsToAdd);
+                }
+
+                // Once added to the definition manager we can update their status
+                await ApplySearchParameterStatus(statuses, cancellationToken);
+
+                var inCache = ParametersAreInCache(statusesToFetch, cancellationToken);
+
+                // if cache is updated directly and not from the database not all will have corresponding resources. Do not advance timestamp as results are not conclusive.
+                if (results.LastUpdated.HasValue && inCache && allHaveResources) // this should be the ony place in the code to assign last updated
+                {
+                    _searchParamLastUpdated = results.LastUpdated.Value;
+                }
+            }
+            finally
+            {
+                try
+                {
+                    _refreshSemaphore.Release();
+                }
+                catch (ObjectDisposedException)
+                {
+                    // Expected during host shutdown when Dispose() races with an in-flight async callback.
+                }
+            }
+
+            return true;
+        }
+
+        private void DeleteSearchParameterIfExists(string url)
+        {
+            try
+            {
+                DeleteSearchParameter(url, false);
+            }
+            catch (ResourceNotFoundException)
+            {
+                // do nothing, there may not be a search parameter to remove
+            }
+        }
+
+        private async Task ApplySearchParameterStatus(IReadOnlyCollection<ResourceSearchParameterStatus> updatedSearchParameterStatus, CancellationToken cancellationToken)
+        {
+            if (!updatedSearchParameterStatus.Any())
+            {
+                _logger.LogDebug("ApplySearchParameterStatus: No search parameter status updates to apply.");
+                return;
+            }
+
+            var updated = new List<SearchParameterInfo>();
+
+            foreach (var paramStatus in updatedSearchParameterStatus)
+            {
+                if (TryGetSearchParameter(paramStatus.Uri.OriginalString, out var param))
+                {
+                    var tempStatus = EvaluateSearchParamStatus(paramStatus);
+
+                    param.IsSearchable = tempStatus.IsSearchable;
+                    param.IsSupported = tempStatus.IsSupported;
+                    param.IsPartiallySupported = tempStatus.IsPartiallySupported;
+                    param.SortStatus = paramStatus.SortStatus;
+                    param.SearchParameterStatus = paramStatus.Status;
+                    updated.Add(param);
+                }
+                else if (!updatedSearchParameterStatus.Any(p => p.Uri.Equals(paramStatus.Uri) && (p.Status == SearchParameterStatus.Deleted || p.Status == SearchParameterStatus.Disabled)))
+                {
+                    // if we cannot find the search parameter in the search parameter definition manager
+                    // and there is an entry in the list of updates with a delete status then it indicates
+                    // the search parameter was deleted before it was added to this instance, and there is no issue
+                    // however if there is no indication that the search parameter was deleted, then there is a problem
+                    _logger.LogError(Core.Resources.UnableToUpdateSearchParameter, paramStatus.Uri);
+                }
+            }
+
+            _statusDataStore.SyncStatuses(updatedSearchParameterStatus);
+
+            _logger.LogDebug("ApplySearchParameterStatus: Synced params. Updated cache timestamp.");
+            await _mediator.Publish(new SearchParametersUpdatedNotification(updated), cancellationToken);
+        }
+
+        private async Task<(IReadOnlyCollection<ResourceSearchParameterStatus> Statuses, DateTimeOffset? LastUpdated)> GetSearchParameterStatusUpdates(CancellationToken cancellationToken, DateTimeOffset? startLastUpdated = null)
+        {
+            var searchParamStatuses = await _statusDataStore.GetSearchParameterStatuses(cancellationToken, startLastUpdated);
+            var lastUpdated = searchParamStatuses.Any() ? searchParamStatuses.Max(_ => _.LastUpdated) : (DateTimeOffset?)null;
+            return (searchParamStatuses, lastUpdated);
+        }
+
+        // This should handle racing condition between saving new parameter on one VM and refreshing cache on the other,
+        // when refresh is invoked between saving status and saving resource.
+        // This will not be needed when order of saves is reversed (resource first, then status)
+        private bool ParametersAreInCache(IReadOnlyCollection<ResourceSearchParameterStatus> statuses, CancellationToken cancellationToken)
+        {
+            var inCache = true;
+            foreach (var status in statuses)
+            {
+                TryGetSearchParameter(status.Uri.OriginalString, out var existingSearchParam);
+                if (existingSearchParam == null)
+                {
+                    var msg = $"Did not find in cache uri={status.Uri.OriginalString} status={status.Status}";
+                    _logger.LogInformation(msg);
+
+                    // if the parameter was updated in the last 10 minutes it's possible we hit race condition
+                    // where status was updated but resource is not yet saved, so we should not consider this as cache miss
+                    if (status.LastUpdated > DateTimeOffset.UtcNow.AddMinutes(-10))
+                    {
+                        inCache = false;
+                    }
+                }
+            }
+
+            return inCache;
+        }
+
+        private async Task<Dictionary<string, ITypedElement>> GetSearchParametersByUrls(List<string> urls, CancellationToken cancellationToken)
+        {
+            if (!urls.Any())
+            {
+                return new Dictionary<string, ITypedElement>();
+            }
+
+            const int chunkSize = 100;
+            var searchParametersByUrl = new Dictionary<string, ITypedElement>();
+
+            // Process URLs in chunks to avoid SQL query limitations
+            for (int i = 0; i < urls.Count; i += chunkSize)
+            {
+                var urlChunk = urls.Skip(i).Take(chunkSize).ToList();
+
+                using IScoped<ISearchService> search = _searchServiceFactory.Invoke();
+
+                // Build a query like: url=url1,url2,url3
+                var urlQueryValue = string.Join(",", urlChunk);
+                var queryParams = new List<Tuple<string, string>>
+                {
+                    new Tuple<string, string>("url", urlQueryValue),
+                    new Tuple<string, string>("_count", chunkSize.ToString()), // we only need a maximum of chunkSize results back
+                };
+
+                var result = await search.Value.SearchAsync(KnownResourceTypes.SearchParameter, queryParams, cancellationToken);
+
+                if (result != null)
+                {
+                    foreach (var searchResultEntry in result.Results)
+                    {
+                        var typedElement = searchResultEntry.Resource.RawResource.ToITypedElement(_modelInfoProvider);
+                        var url = typedElement.GetStringScalar("url");
+
+                        if (!string.IsNullOrEmpty(url))
+                        {
+                            if (!searchParametersByUrl.ContainsKey(url))
+                            {
+                                searchParametersByUrl[url] = typedElement;
+                            }
+                            else
+                            {
+                                _logger.LogWarning("More than one SearchParameter found with url {Url}. Using the first one found.", url);
+                            }
+                        }
+                    }
+                }
+            }
+
+            return searchParametersByUrl;
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                _refreshSemaphore?.Dispose();
+            }
+        }
+
+        private static TempStatus EvaluateSearchParamStatus(ResourceSearchParameterStatus paramStatus)
+        {
+            TempStatus tempStatus;
+            tempStatus.IsSearchable = paramStatus.Status == SearchParameterStatus.Enabled;
+            tempStatus.IsSupported = paramStatus.Status == SearchParameterStatus.Supported || paramStatus.Status == SearchParameterStatus.Enabled;
+            tempStatus.IsPartiallySupported = paramStatus.IsPartiallySupported;
+
+            return tempStatus;
+        }
+
+        private struct TempStatus
+        {
+            public bool IsSearchable;
+            public bool IsSupported;
+            public bool IsPartiallySupported;
         }
     }
 }
